sys-requirements.txt
# SPDX-FileCopyrightText: 2019-2020 Magenta ApS
# SPDX-License-Identifier: MPL-2.0

# This file contains system package dependencies. It is used for `apt-get` in
# Dockerfile and .gitlab-ci.yml. It is intended for the python:3.6 docker image
# which through buildpack-deps is based on Debian Buster.

# Python packages dependencies:
# for xmlsec.
libxmlsec1-dev
